+++ src/components/common/Score/Score.styles.ts
@@ -5,12 +5,12 @@ export default createUseStyles((theme: ITheme) => ({
   input: {
     backgroundColor: `${theme.inputBackgroundColor} !important`,
     color: `${theme.inputTextColor} !important`,
-    fontSize: "1.5rem",
+    fontSize: "1.25rem",
     height: "100%",
-    padding: "0.25rem",
+    padding: "0 0.5rem",
     textAlign: "center",
     appearance: "textfield !important",
     "-moz-appearance": "textfield !important",
-    width: "2rem",
+    width: "2.5rem",
   },
 }));