+++ src/context/index.tsx
@@ -13,7 +13,7 @@ export interface IDataProviderProps {
 }
 
 export const INITIAL_DATA: IData = {
-  version: "0.8.3",
+  version: "0.9.0",
   theme: "dark",
   innocentWins: 0,
   innocentLosses: 0,